Protected route (#3)
* feat: new auth middleware

* feat: admin and workers protected routes for users

+++ src/controllers/UserController.js
@@ -1,116 +1,51 @@
 const express = require('express');
 const { User } = require('../models/UserModel');
 const router = express.Router();
-const { comparePassword, generateJwt } = require('../functions/userAuthFunctions');
-// const { validJWT, catchErrors}= require('../middleware/admin_auth')
+const { authUser } = require('../middleware/admin_auth');
+const bcrypt = require("bcryptjs");
+const jwt = require('jsonwebtoken');
+require('dotenv').config();
+
+
 
 // get all users
-// router.get('/all',  async (request, response) => {
-		// let decodedJWT = request.decodedJWT;
-	
-		// // Assign locally for cleaner code - but not needed.
-		// let decryptedEmail = decryptString(decodedJWT.email);
-		// let decryptedPassword = decryptString(decodedJWT.password);
-	
-		// // To prove we got the data we're expecting:
-		// console.log("Decrypted user data is: \n\t" + decryptedEmail + "\n\t" + decryptedPassword);
-	
-		// // Encrypt the data again to make sure nothing unsecure is put into the new, fresh JWT
-		// let encryptedEmail = encryptString(decryptedEmail);
-		// let encryptedPassword = encryptString(decryptedPassword);
-	
-		// let objectToTokenize = {
-		// 	email: encryptedEmail, 
-		// 	password: encryptedPassword
-		// }
-	
-
-		// // Passing in an object instead of doing:
-		// // generateJWT({someKey:"someValue"})
-		// // to make that above if condition easier to write
-		// let userJwt = generateJWT(objectToTokenize);
-	
-		// // The response can assume the golden happy path,
-		// // since error-handling middleware would send its own response
-		// // if an error were detected.
-		// let responseData = {
-		// 	message: "You are a valid user!",
-		// 	newJWT: userJwt
-		// }
-	// 	let result = await User.find({}).populate('role', 'name');
-	// 	response.json(responseData);
-	// response.json({
-	// 	user: result
-	// });
-	// });
-
-  
-  // GET all users.
-  router.get('/all',  async (request, response) => {
-	let result = await User.find({}).populate('role', 'name');
-	response.json({
-		user: result
-	});
-
+// Admin and workers only
+router.get('/all', authUser,  async (request, response) => {
+	if (request.user.role === 'admin'|| 'worker'){
+		let result = await User.find({}).populate('role', 'name').sort({ createdAt: -1 });
+			response.json({user: result})
+	} else{
+		response.status(403).json({error: 'Access Forbidden'})
+	}
 });
 
+
 // find one user by id
-router.get("/one/:id",  async (request, response) => {
-	let result = await User.findOne({_id: request.params.id});
-	// response.json({ message: 'This route is accessible only by admins.' });
-	response.json({
-		user: result
-	});
+// Admin and workers only
+router.get("/one/:id", authUser, async (request, response) => {
+	if (request.user.role === 'admin' || 'worker'){
+		let result = await User.findOne({_id: request.params.id});
+		response.json({user: result})
+	} else{
+		response.status(403).json({error: 'Access Forbidden'})
+	}
 });
 
 // find user by last name
+// Admin and workers only
 router.get("/name/:lastName", async (request, response) => {
-	let result = await User.find({lastName: request.params.lastName});
-
-	response.json({
-		user: result
-	});
-});
-
-
-
-// create new user
-router.post("/", async (request, response) => {
-	let newUser = await User.create(request.body).catch(error => error);
-
-    response.json({
-		user: newUser
-	});
-});
-
-
-// POST to /users/login
-router.post("/login", async (request, response) => {
-	// Find user by provided email 
-	let targetUser = await User.findOne({email: request.body.email}).catch(error => error);
-
-	// Check if user provided the correct password
-	let isPasswordCorrect = await comparePassword(request.body.password, targetUser.password);
-
-	if (!isPasswordCorrect){
-		response.status(403).json({error:"Invalid Details, Try Again!"});
+	if (request.user.role === 'admin'|| 'worker'){
+		let result = await User.find({lastName: request.params.lastName});
+	} else{
+		response.status(403).json({error: 'Access Forbidden'})
 	}
-
-	// If they provided the correct, generate a JWT
-	let freshJwt = generateJwt(targetUser._id.toString());
-
-	// respond with the JWT 
-	response.json({
-		jwt: freshJwt
 	});
-
-});
-
-
 
 // update user
 // does not overwrite or remove any unmentioned properties
-router.patch("/:id", async (request, response) => {
+// Admin and workers only
+router.patch("/:id",authUser, async (request, response) => {
+	if (request.user.role === 'admin' || 'worker'){
 	let result = await User.findByIdAndUpdate(
 		request.params.id, 
 		request.body,
@@ -120,21 +55,65 @@ router.patch("/:id", async (request, response) => {
 		}
 		).catch(error => error);
 
-	response.json({
-		user: result
-	});
+	response.json({user: result});
 
+} else{
+	response.status(403).json({error: 'Access Forbidden'})
+}
 });
 
 
-// delete user by id
-router.delete("/:id", async (request, response) => {
-	let result = await User.findByIdAndDelete(request.params.id).catch(error => error);
-    
-	response.json({
-		deletedUser: result
-	});
-
+// delete user by id. 
+// Admin only
+router.delete("/delete/:id",authUser, async (request, response) => {
+	if (request.user.role === 'admin'){
+		let result = await User.findByIdAndDelete(request.params.id).catch(error => error);
+		response.json({deletedUser: result});
+	} else{
+		response.status(403).json({error: 'Access Forbidden'})
+	}
 });
+
+// create new user
+router.post("/register", async (request, response) => {
+	let newUser = await User.create(request.body).catch(error => error);
+
+	response.status(201).json({message: 'User registered successfully!'});
+
+    response.json({
+		user: newUser
+	});
+});
+
+
+// POST to /users/login
+router.post("/login", async (request, response) => {
+
+	// Find user by provided email 
+	const { email, password } = request.body;
+
+	try{
+		const user = await User.findOne({ email });
+		
+		if (!user){
+			return response.status(401).json({message: 'Invalid Email or Password!'})
+		}
+
+		const passwordMatch = await bcrypt.compare(password, user.password);
+
+		if (!passwordMatch){
+			return response.status(401).json({message: 'Invalid Email or Password!'})
+		}
+
+		const token = jwt.sign({id: user._id, email: user.email, role: user.role}, process.env.USER_JWT_KEY, { expiresIn: '3h'});
+
+		response.json({token});
+	}
+	catch (error) {
+		response.status(500).json({error: error.message});
+	}
+});
+
+
 
 module.exports = router;
